+++ src/asp/Sessions/StereoSessionNadirPinhole.cc
@@ -77,37 +77,13 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
                                       this->m_out_prefix, right_cropped_file);
 
   // TODO(oalexan1): Should the masks be uint8 instead of float?
-  ImageViewRef<PixelMask<float>> left_bathy_mask, right_bathy_mask;
-  bool do_bathy = StereoSession::do_bathymetry();
-  float left_bathy_nodata = -std::numeric_limits<float>::max();
-  float right_bathy_nodata = -std::numeric_limits<float>::max();
-  if (do_bathy) 
-    StereoSession::read_bathy_masks(left_masked_image, right_masked_image,
-                                    left_bathy_nodata, right_bathy_nodata,
-                                    left_bathy_mask, right_bathy_mask);
-  
   ImageViewRef<PixelMask<float>> Limg, Rimg;
-  ImageViewRef<PixelMask<float>> left_aligned_bathy_mask, right_aligned_bathy_mask;
 
   // Use no-data in interpolation and edge extension.
   PixelMask<float>nodata_pix(0); nodata_pix.invalidate();
-  PixelMask<float>bathy_nodata_pix(0); bathy_nodata_pix.invalidate();
   ValueEdgeExtension<PixelMask<float>> ext_nodata(nodata_pix); 
-  ValueEdgeExtension<PixelMask<float>> bathy_ext_nodata(bathy_nodata_pix); 
   
   if ( stereo_settings().alignment_method == "epipolar" ) {
-
-    if (do_bathy) {
-      // I could not code and test in reasonable time bathymery with cropped
-      // images and epipolar alignment. 
-      bool crop_left  = (stereo_settings().left_image_crop_win  != BBox2i(0, 0, 0, 0));
-      bool crop_right = (stereo_settings().right_image_crop_win != BBox2i(0, 0, 0, 0));
-      if (crop_left || crop_right) 
-        vw_throw(NoImplErr() << "Bathymetry computation is not supported with epipolar "
-                 << "alignment while using --left-image-crop-win or --right-image-crop-win. "
-                 << "Use the full images.");
-    }
-    
     vw_out() << "\t--> Performing epipolar alignment\n";
 
     // Load the two images and fetch the two camera models
@@ -145,19 +121,6 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
                                               Limg, Rimg,
                                               ext_nodata,
                                               BilinearInterpolation());
-
-      if (do_bathy)
-        get_epipolar_transformed_pinhole_images(m_left_camera_file, m_right_camera_file,
-                                                left_cam, right_cam,
-                                                left_bathy_mask, right_bathy_mask,
-                                                left_image_in_roi, right_image_in_roi,
-                                                left_out_size, right_out_size,
-                                                left_aligned_bathy_mask,
-                                                right_aligned_bathy_mask,
-                                                bathy_ext_nodata,
-                                                BilinearInterpolation());
-        
-        
     } else { // Handle CAHV derived models
 
       camera_models(left_cam, right_cam);
@@ -166,14 +129,6 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
                                       left_cam, right_cam,
                                       left_masked_image, right_masked_image,
                                       Limg, Rimg, ext_nodata);
-
-      if (do_bathy)
-        get_epipolar_transformed_images(m_left_camera_file, m_right_camera_file,
-                                        left_cam, right_cam,
-                                        left_bathy_mask, right_bathy_mask,
-                                        left_aligned_bathy_mask,
-                                        right_aligned_bathy_mask,
-                                        bathy_ext_nodata);
     }
     
   } else if (stereo_settings().alignment_method == "homography"     ||
@@ -239,28 +194,15 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
     Limg = transform(left_masked_image,
                      HomographyTransform(align_left_matrix),
                      left_size.x(), left_size.y() );
-    if (do_bathy) 
-      left_aligned_bathy_mask = transform(left_bathy_mask,
-                                          HomographyTransform(align_left_matrix),
-                                          left_size.x(), left_size.y() );
-    
     
     Rimg = transform(right_masked_image,
                      HomographyTransform(align_right_matrix),
                      right_size.x(), right_size.y() );
-    if (do_bathy) 
-      right_aligned_bathy_mask = transform(right_bathy_mask,
-                                           HomographyTransform(align_right_matrix),
-                                           right_size.x(), right_size.y() );
     
   } else {
     // Do nothing just provide the original files.
     Limg = left_masked_image;
     Rimg = right_masked_image;
-    if (do_bathy) {
-      left_aligned_bathy_mask = left_bathy_mask;
-      right_aligned_bathy_mask = right_bathy_mask;
-    }
   }
   
   // Apply our normalization options.
@@ -297,17 +239,6 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
                           has_nodata, output_nodata,
                           options,
                           TerminalProgressCallback("asp","\t  L:  ") );
-  if (do_bathy) {
-    std::string left_aligned_bathy_mask_file = StereoSession::left_aligned_bathy_mask();
-    vw_out() << "\t--> Writing: " << left_aligned_bathy_mask_file << ".\n";
-    block_write_gdal_image(left_aligned_bathy_mask_file, apply_mask(left_aligned_bathy_mask,
-                                                                    left_bathy_nodata),
-                           has_left_georef, left_georef,
-                           has_bathy_nodata, left_bathy_nodata,
-                           options,
-                           TerminalProgressCallback("asp","\t  L mask:  ") );
-  }
-
   vw_out() << "\t--> Writing: " << right_output_file << ".\n";
   block_write_gdal_image(right_output_file,
                          // Force R.tif to be the same size as L.tif.
@@ -317,15 +248,8 @@ void asp::StereoSessionNadirPinhole::pre_preprocessing_hook(bool adjust_left_ima
                          has_right_georef, right_georef,
                          has_nodata, output_nodata, options,
                          TerminalProgressCallback("asp","\t  R:  ") );
-  if (do_bathy) {
-    std::string right_aligned_bathy_mask_file = StereoSession::right_aligned_bathy_mask();
-    vw_out() << "\t--> Writing: " << right_aligned_bathy_mask_file << ".\n";
-    block_write_gdal_image(right_aligned_bathy_mask_file,
-                           apply_mask(crop(edge_extend(right_aligned_bathy_mask, bathy_ext_nodata), 
-                                           bounding_box(Limg)), right_bathy_nodata),
-                           has_right_georef, right_georef,
-                           has_bathy_nodata, right_bathy_nodata,
-                           options,
-                           TerminalProgressCallback("asp","\t  R bathy mask:  ") );
-  }
+
+  if (this->do_bathymetry())
+    this->align_bathy_masks(options);
+
 }